fix: bypass nextjs router convention by rensminf routes.ts to index.ts
